Refine cookie consent message
