+++ assets/js/bootstraps/common.js
@@ -8,7 +8,7 @@ import heroImages from '../modules/heroImages';
 const logos = require('../modules/logos');
 import forms from '../modules/forms';
 import carousel from '../modules/carousel';
-import cookieWarning from '../modules/cookieWarning';
+import cookieConsent from '../modules/cookie-consent';
 
 function init() {
     global.init();
@@ -17,7 +17,7 @@ function init() {
     logos.init();
     forms.init();
     carousel.init();
-    cookieWarning.init();
+    cookieConsent.init();
 }
 
 export default {
