+++ assets/sass/style.scss
@@ -62,7 +62,6 @@
 @import 'components/news';
 @import 'components/global-header';
 @import 'components/global-footer';
-@import 'components/cookie-warning';
 
 // =========================================================================
 // Utilities
