+++ views/includes/cookieWarning.njk
@@ -1,5 +1,14 @@
-<aside class="cookie-warning u-hidden" id="js-cookie-warning">
-    <h4>Cookies on this website</h4>
-    <p>We use cookies to give you the best online experience. Please let us know if you agree to all of these cookies. <a href="{{ localify('/about/customer-service/cookies') }}">More information on how we use cookies</a>.</p>
-    <button class="btn btn--small" id="js-accept-cookies">Accept cookies</button>
+<aside class="cookie-consent" id="js-cookie-consent">
+    <div class="cookie-consent__inner">
+        <div class="cookie-consent__content">
+            <h4 class="cookie-consent__title">Cookies on this website</h4>
+            <div class="cookie-consent__message">
+                <p>This site uses cookies.  An explanation of their purpose can be found within our <a href="{{ localify('/about/customer-service/cookies') }}">Cookies Policy</a>.</p>
+                <p>Please confirm your consent to their use by clicking continue.  After consenting you will not see this message again.</p>
+            </div>
+        </div>
+        <div class="cookie-consent__actions">
+            <button class="btn btn--small" id="js-cookie-consent-accept">Accept cookies</button>
+        </div>
+    </div>
 </aside>
